+++ src/material.cpp
@@ -33,7 +33,7 @@ void gl::Material::Bind() const
 }
 void gl::Material::Unbind()
 {
-    for (size_t i = 0; i < 4; i++)
+    for (int i = 0; i < 4; i++)
     {
         glActiveTexture(GL_TEXTURE0 + i);
         glBindTexture(GL_TEXTURE_2D, 0);
